src/webapp/requirements.txt
alembic==0.8.7
bcrypt==3.1.0
cffi==1.7.0
click==6.6
coverage==4.2
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.5.0
GitHub-Flask==3.1.3
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
py==1.4.31
py-bcrypt==0.4
pycparser==2.14
pytest==3.0.1
pytest-cov==2.3.1
pytest-flask==0.10.0
python-editor==1.0.1
requests==2.11.1
six==1.10.0
Werkzeug==0.11.10
